Fix/113 홈화면 확정된 약속 조회 날짜 형식 수정

+++ src/main/java/com/kuit/conet/dto/web/response/team/GetTeamResponseDTO.java
@@ -16,11 +16,11 @@ public class GetTeamResponseDTO {
     private Boolean bookmark;
 
 
-    public GetTeamResponseDTO(Team team, Long teamMemberCount, Boolean bookmark) {
+    public GetTeamResponseDTO(Team team, int teamMemberCount, Boolean bookmark) {
         this.teamId = team.getId();
         this.teamName = team.getName();
         this.teamImgUrl = team.getImgUrl();
-        this.teamMemberCount = teamMemberCount;
+        this.teamMemberCount = Long.valueOf(teamMemberCount);
         this.bookmark = bookmark;
 
         //isNew 결정
